refactor: 파일 이동 (auth/token/jwt/ -> auth/jwt/)
