--- backend/src/main/java/project/backend/auth/token/dao/TokenRedisRepository.java
@@ -1,8 +1,0 @@
-package project.backend.auth.token.dao;
-
-import org.springframework.data.repository.CrudRepository;
-import project.backend.auth.token.entity.TokenRedis;
-
-public interface TokenRedisRepository extends CrudRepository<TokenRedis, Long> {
-
-}
