+++ backend/src/main/java/project/backend/global/security/entrypoint/RestAuthenticationEntryPoint.java
@@ -8,8 +8,6 @@ import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
-import project.backend.global.exception.errorcode.ErrorCode;
-import project.backend.global.exception.ex.CustomJwtException;
 
 @Slf4j
 @Component
@@ -17,23 +15,11 @@ public class RestAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
 	@Override
 	public void commence(HttpServletRequest request, HttpServletResponse response,
-		AuthenticationException authException) throws IOException {
+						 AuthenticationException authException) throws IOException {
 		log.warn("인증되지 않은 사용자 접근: {} - {}", request.getRequestURI(), authException.getMessage());
 
 		response.setContentType("application/json; charset=utf-8");
-
-		String message = null;
-
-		if (authException instanceof CustomJwtException jwtEx) {
-			ErrorCode errorCode = jwtEx.getErrorCode();
-			response.setStatus(errorCode.getStatus().value());
-			message = errorCode.getMessage();
-		} else {
-			response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-			message = "로그인이 필요한 서비스입니다.";
-		}
-
-		response.getWriter()
-			.write("{\"message\":\"" + message + "\"}");
+		response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+		response.getWriter().write("{\"message\":\"로그인이 필요한 서비스입니다.\"}");
 	}
 }
