+++ backend/src/main/java/project/backend/global/security/filter/JwtAuthenticationFilter.java
@@ -13,9 +13,9 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-import project.backend.auth.app.CookieUtils;
-import project.backend.auth.token.jwt.JwtProvider;
-import project.backend.auth.token.jwt.TokenStatus;
+import project.backend.global.util.CookieUtils;
+import project.backend.auth.jwt.JwtProvider;
+import project.backend.auth.jwt.TokenStatus;
 import project.backend.global.exception.errorcode.TokenErrorCode;
 import project.backend.global.exception.ex.CustomJwtException;
 
@@ -27,40 +27,38 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	private final JwtProvider jwtProvider;
 
 	private static final List<String> WHITE_LIST = List.of(
-		"/signup",
-		"/login",
-		"/token/refresh",
-		"/actuator/health",
-		"/actuator/prometheus",
-		"/ws"
+			"/signup",
+			"/login",
+			"/token/refresh",
+			"/actuator/health",
+			"/actuator/prometheus",
+			"/ws"
 	);
 
 	@Override
 	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
-		FilterChain filterChain)
-		throws ServletException, IOException {
+									FilterChain filterChain) throws ServletException, IOException {
 
 		String requestURI = request.getRequestURI();
-		log.debug("[JWT Filter] 요청 URI: {}", requestURI);
-		if (requestURI.startsWith("/github/") || WHITE_LIST.contains(request.getRequestURI())) {
-			filterChain.doFilter(request, response); // JWT 검사 건너뜀
+		log.info("[JWT Filter] 요청 URI: {}", requestURI);
+
+		if (requestURI.startsWith("/github/") || WHITE_LIST.contains(requestURI)) {
+			filterChain.doFilter(request, response);
 			return;
 		}
 
-		log.info("JWT필터 도달 = {}", request.getRequestURI());
+		log.info("JWT필터 도달 = {}", requestURI);
 
-		Cookie cookie = CookieUtils.getCookie(request,
-				"accessToken")
-			.orElseThrow(() -> new CustomJwtException(TokenErrorCode.NOT_FOUND_TOKEN));
+		Cookie cookie = CookieUtils.getCookie(request, "accessToken")
+				.orElseThrow(() -> new CustomJwtException(TokenErrorCode.NOT_FOUND_TOKEN));
+
 		String accessToken = cookie.getValue();
-
 		TokenStatus tokenStatus = jwtProvider.validateAccessToken(accessToken);
 
 		if (tokenStatus == TokenStatus.VALID) {
 			log.info("[JWT] 유효한 토큰");
 			Authentication authentication = jwtProvider.getAuthentication(accessToken);
 			SecurityContextHolder.getContext().setAuthentication(authentication);
-
 		} else {
 			SecurityContextHolder.clearContext();
 		}
